requirements.txt
streamlit
google-generativeai
youtube-transcript-api
fpdf
requests
python-dotenv
PyPDF2
matplotlib
numpy
langchain
faiss-cpu
langchain-google-genai
